// src/app/store/model/report.ts
export interface Report {
  reportId: string;
  reportedUUID: string;
  reportUUID: string;
  reportedName: string;
  reportName: string;
  reportTime: string;
  reason: string;
  replay: string;
  server: string;
  assignedUser: string;
  selectedSystem: string;
  selectedReason: string;
  reportState: string;
  rank: string;
  onlineState: string;
}